requirements.txt
langchain_text_splitters==0.3.0
loguru==0.7.2
numpy==2.1.1
openai==1.48.0
pandas==2.2.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pytube==15.0.0
scikit_learn==1.5.2
streamlit==1.38.0
youtube_transcript_api==0.6.2
